A classic UNIX forkbomb with many extensions
A classic UNIX forkbomb with various file extensions. **WARNING: This file is a forkbomb. When ran, it will duplicate itself over and over until your system is out of memory. Modern *NIX systems have systems in place to prevent this, but you should still exercise caution. Try running it in a Virtual Machine first.